fix(role balance type): Convert bigint to text type for role.balance

+++ libs/dao/src/entities/role.entity.ts
@@ -31,7 +31,7 @@ export class Role extends BaseEntity {
   kind: RoleKindType;
 
   @ApiProperty()
-  @Column({ type: 'bigint', nullable: true })
+  @Column({ type: 'numeric', nullable: true })
   balance: number;
 
   @ApiProperty()
